Rebased #178 on master
