+++ remove_artifacts.sh
@@ -10,7 +10,11 @@ artifacts="
   buildaux/
   configure
   include/Makefile.in
-  m4/
+  m4/libtool.m4
+  m4/lt~obsolete.m4
+  m4/ltoptions.m4
+  m4/ltsugar.m4
+  m4/ltversion.m4
   src/Makefile.in
   src/lib/Makefile.in
   src/lib/config.h.in
