+++ src/cmocka/rnp_tests.h
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2017, [Ribose Inc](https://www.ribose.com).
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
+ *
+ * 1.  Redistributions of source code must retain the above copyright notice,
+ *     this list of conditions and the following disclaimer.
+ *
+ * 2.  Redistributions in binary form must reproduce the above copyright notice,
+ *     this list of conditions and the following disclaimer in the documentation
+ *     and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+ * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+void rnpkeys_exportkey_verifyUserId(void **state);
+
+void rnpkeys_generatekey_testSignature(void **state);
+
+void rnpkeys_generatekey_testEncryption(void **state);
+
+void rnpkeys_generatekey_verifySupportedHashAlg(void **state);
+
+void rnpkeys_generatekey_verifyUserIdOption(void **state);
+
+void rnpkeys_generatekey_verifykeyHomeDirOption(void **state);
+
+void rnpkeys_generatekey_verifykeyNonexistingHomeDir(void **state);
+
+void rnpkeys_generatekey_verifykeyHomeDirNoPermission(void **state);
+
+void rnpkeys_exportkey_verifyUserId(void **state);
+
+void hash_test_success(void **state);
+
+void cipher_test_success(void **state);
+
+void pkcs1_rsa_test_success(void **state);
+
+void raw_elg_test_success(void **state);
