+++ src/lib/crypto.h
@@ -78,8 +78,8 @@ unsigned pgp_dsa_verify(const uint8_t *,
                         const pgp_dsa_pubkey_t *);
 
 /*
-* RSA encrypt/decrypt
-*/
+ * RSA encrypt/decrypt
+ */
 
 int pgp_rsa_encrypt_pkcs1(uint8_t *               out,
                           size_t                  out_len,
@@ -95,12 +95,12 @@ int pgp_rsa_decrypt_pkcs1(uint8_t *               out,
                           const pgp_rsa_pubkey_t *pubkey);
 
 /*
-* RSA signature generation and verification
-*/
+ * RSA signature generation and verification
+ */
 
 /*
-* Returns 1 for valid 0 for invalid/error
-*/
+ * Returns 1 for valid 0 for invalid/error
+ */
 int pgp_rsa_pkcs1_verify_hash(const uint8_t *         sig_buf,
                               size_t                  sig_buf_size,
                               pgp_hash_alg_t          hash_alg,
@@ -109,8 +109,8 @@ int pgp_rsa_pkcs1_verify_hash(const uint8_t *         sig_buf,
                               const pgp_rsa_pubkey_t *pubkey);
 
 /*
-* Returns # bytes written to sig_buf on success, 0 on error
-*/
+ * Returns # bytes written to sig_buf on success, 0 on error
+ */
 int pgp_rsa_pkcs1_sign_hash(uint8_t *      sig_buf,
                             size_t         sig_buf_size,
                             pgp_hash_alg_t hash_alg,
@@ -120,21 +120,21 @@ int pgp_rsa_pkcs1_sign_hash(uint8_t *      sig_buf,
                             const pgp_rsa_pubkey_t *);
 
 /*
-* Performs ElGamal encryption
-* Result of an encryption is composed of two parts - g2k and encm
-*
-* @param g2k [out] buffer stores first part of encryption (g^k % p)
-* @param encm [out] buffer stores second part of encryption (y^k * in % p)
-* @param in plaintext to be encrypted
-* @param length length of an input
-* @param pubkey public key to be used for encryption
-*
-* @pre g2k size must be at least equal to byte size of prime `p'
-* @pre encm size must be at least equal to byte size of prime `p'
-*
-* @return 	on success - number of bytes written to g2k and encm
-*			on failure -1
-*/
+ * Performs ElGamal encryption
+ * Result of an encryption is composed of two parts - g2k and encm
+ *
+ * @param g2k [out] buffer stores first part of encryption (g^k % p)
+ * @param encm [out] buffer stores second part of encryption (y^k * in % p)
+ * @param in plaintext to be encrypted
+ * @param length length of an input
+ * @param pubkey public key to be used for encryption
+ *
+ * @pre g2k size must be at least equal to byte size of prime `p'
+ * @pre encm size must be at least equal to byte size of prime `p'
+ *
+ * @return 	on success - number of bytes written to g2k and encm
+ *			on failure -1
+ */
 int pgp_elgamal_public_encrypt_pkcs1(uint8_t *                   g2k,
                                      uint8_t *                   encm,
                                      const uint8_t *             in,
@@ -142,22 +142,22 @@ int pgp_elgamal_public_encrypt_pkcs1(uint8_t *                   g2k,
                                      const pgp_elgamal_pubkey_t *pubkey);
 
 /*
-* Performs ElGamal decryption
-*
-* @param out [out] decrypted plaintext
-* @param g2k buffer stores first part of encryption (g^k % p)
-* @param encm buffer stores second part of encryption (y^k * in % p)
-* @param length length of g2k or in (must be equal to byte size of prime `p')
-* @param seckey private part of a key used for decryption
-* @param pubkey public domain parameters (p,g) used for decryption
-*
-* @pre g2k size must be at least equal to byte size of prime `p'
-* @pre encm size must be at least equal to byte size of prime `p'
-* @pre byte-size of `g2k' must be equal to `encm'
-*
-* @return 	on success - number of bytes written to g2k and encm
-*			on failure -1
-*/
+ * Performs ElGamal decryption
+ *
+ * @param out [out] decrypted plaintext
+ * @param g2k buffer stores first part of encryption (g^k % p)
+ * @param encm buffer stores second part of encryption (y^k * in % p)
+ * @param length length of g2k or in (must be equal to byte size of prime `p')
+ * @param seckey private part of a key used for decryption
+ * @param pubkey public domain parameters (p,g) used for decryption
+ *
+ * @pre g2k size must be at least equal to byte size of prime `p'
+ * @pre encm size must be at least equal to byte size of prime `p'
+ * @pre byte-size of `g2k' must be equal to `encm'
+ *
+ * @return 	on success - number of bytes written to g2k and encm
+ *			on failure -1
+ */
 int pgp_elgamal_private_decrypt_pkcs1(uint8_t *                   out,
                                       const uint8_t *             g2k,
                                       const uint8_t *             in,
@@ -332,6 +332,6 @@ struct pgp_stream_t {
  * \brief Allocates BIGNUM and mp value assigned
  */
 BIGNUM *new_BN_take_mp(botan_mp_t mp);
-void destroy_BN_mp(BIGNUM** a);
+void destroy_BN_mp(BIGNUM **a);
 
 #endif /* CRYPTO_H_ */
